requirements.txt
fastapi
uvicorn
google-genai
firebase-admin
pydantic
python-dotenv
